Simple interest Calculation

+++ python_introduction/simple_interest.py
@@ -4,10 +4,10 @@ rate = 0.05
 time = 3
 
 # Making Sure Variables are intergers
-interest = int(principal * rate * time)
+interest = principal * rate * time
 principal = int(1000.0)
 rate = int(0.050)
 time = int(3.0)
 
 #Performing Calculations and printing their results
-print('The simple interest is:', interest)
+print("The simple interest is:", interest)
